Factor out `completion` and eliminate `types`
